requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.0.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pkg_resources==0.0.0
psycopg2==2.9.3
python-dotenv==0.17.1
pytz==2021.1
PyYAML==6.0
sqlparse==0.4.1
tablib==3.2.1
typing-extensions==3.10.0.0
xlrd==2.0.1
xlwt==1.3.0
